moving playback related handlers to separate file
